add makeProject unit tests
